Update will prep for new minting functions
## Contract Documentation
The contract's name was changed to reflect the current project. Todos
adjusted to show what was currently being worked on. A Minting Function
section was added for organizational purposes.  The collection name and
symbol were also updated.

## Image Struct and Mapping
Because there are 50 images with varying properties it didn't make
sense to have a minting function for each image which was the previous
solution for the PuppyMill. Instead  a struct was created. It stores
the images' id, price and  supplyLimit. A 4th variable was added as
padding so that the struct packages compile into a single  container
when deployed to help save gas. A mapping was used to retrieve the
information from the structs. Mappings cut gas in half compared to
arrays in my testing.

## Minting Functions Removed
Two of the three minting functions were removed, as they are no longer
necessary. The plan is to pass an image ID to the function from the
front end, so it mints the correct image. The struct and minting
functions will need to be adjusted once the problem is solved. I am
still working on it in my head.