Add new minting function and imageGallery mapping
## Image Struct Changes
The Image Struct has been updated to store the following, imgNumber,
price, currentTokenId and lastTokenId.

## Initializes the imageGallery
The imageGallery stores all the Image struct objects. It is
initialized in the constructor and specifies the details for each
image including the tokenIds which will point to the correct metadata
for each image. A list of the tokensId for each image are stored in
./documentation/TokenIdRangesByImageNumber.txt for reference and
testing.

## New Minting Function
The new minting function was added mintImage(). It requries a
_imageNumber to be passed and will mint a token for that image as long
as there are tokens left. Returns the newTokenId.