Add owner functionality to the smart contract
## OpenZeppelin's Ownable.sol
The Ownable.sol contract has been imported providing support for an
'owner' of the contract to be specified. More information on this
extension can be found at
https://docs.openzeppelin.com/contracts/4.x/access-control

## Additional Functionality
The 'owner' of this contract is set as the wallet address which
deployed the contract via the constructor.

The onlyOwner modifier can now be added to functions and will prevent
anyone but the owner from calling it.

Ownership can be transferred with the transferOwnership() function.

The renounceOwnership() will set the Owner to the zero address and
prevent onlyOwner functions from being called. This is permanent!!