Add royalty functionality and extension
## OpenZeppelin's ERC721Royalty.sol
The ERC721Royalty.sol contract extension has been imported providing
support for setting royalties via the ERC-2981 specifications. More
information on this extension can be found at
https://docs.openzeppelin.com/contracts/4.x/api/token/erc721#ERC721Royalty

## ERC721Royalty Extension Set up
The extension was imported and the contract was set to inherit the
extension functionality.

_burn() and supportsInterface() functions were added. Both are required
for the royalty extension.

The dummyMoneypipeAddress variable was created which will be replaced
at a later time with the correct addresses for royalty payouts.

## Minting Function Changes
The _setTokenRoyalty() function is called after _safeMint() completes.
It sets the royalty for the token that was just created. It sets the
payout to the dummyMoneypipeAddress and the percentage to 10%
(1000 basis points).

## Documentation Changes
Modification History Updated
Tasks updated and reorganized.
Additional headings added.
Heading names changed to be more explicit.