Release version 0.1.0 alpha
## 0.1.0 alpha Release
This is the initial contract release for this project. It has changed
significantly from ScottsPuppyMill.sol and felt like a good place to
create a release. There are no major changes from the previous
version.

## Current Functionality
Contract is deployable with a minting function and royalty support
via the ERC-2981 specifications.

## Next Steps
Minting Function Gas Optimization
Charge fee for minting a token.