Add payouts and update to version 0.2.0 alpha
# Version 0.2.0 alpha Changes
Very happy to release RenneLaneCollection.sol version 0.2.0 alpha.

This version supports minting and royalty payouts and payout
splitting via the use of MoneyPipe (MoneyPipe.xyz) contracts. Minting
payouts are made automatically upon successful minting of a token.
Royalty information is created per token and stored as part of the
token following the ERC-2981 standard. When a royalty payment is made
it is directed to a MoneyPipe Contract which splits the amount and
disperses the funds immediately.

!! Please note that the ERC-2981 Royalty Standard is a declaration of
preferences. It does not guarantee royalty payments will be made,
this relies on a market places' honoring of the standard. OpenSea
does not currently honor this standard, but the MoneyPipe Contract
Address and be added manually to OpenSea as well other marketplaces.
It is highly recommended you do this on all marketplaces you expect
your NFTs to be sold. !!

## Minting and Royalty Payouts and Splits
MoneyPipe Contracts handle the splits. Take a look at the design doc
(https://bit.ly/rl_collab) for rational and additional information.

MoneyPipe Contracts needs to be set up separately at (MoneyPipe.xyz).
The design doc (https://bit.ly/rl_collab) has more information.
The Minting Payout Split Contract will send 10% to the Artist and 90%
to the Project Wallet. The Royalty Payout Split Contract will do a
50/50 split between the artist and the Project Wallet.

### Image Struct Changes
The Image Struct has two new variables attached (address
mintingPayoutAddress and address royaltyPayoutAddress) where the
MoneyPipe Contract addresses can be set during the constructor.

!! The current addresses are dummy addresses and will not work.
!! Change them for your deployment or your Ether will be lost.

## Minting Function Changes
mintImage() now sets the royaltyPayoutAddress from the Image Struct.
It will also send the msg.value paid for minting the image to the
mintingPayoutAddress. The docstring for mintImage() was updated to
include the new functionality.

## Assistant Functions Added
These functions allow for information to be pulled from the contract
for testing and verification purposes. Unlikely these will make it to
the production version but they are helpful during development.
getImageInfo() will return information based on an _imageNumber.
getContractStoredBalance() returns the amount of Ether stored inside
the contract. This should usually be zero.

## Documentation Changes
The Statistics, Modification History and Task List were updated for
the current version.