Add Investor List and update to 0.3.0 alpha
# Version 0.3.0 alpha Changes
Very happy to release RenneLaneCollection.sol version 0.3.0 alpha.

This version adds functionality for a permanent investor list which
will save the address of anyone who mints an item from this
collection permanently. This list can be queried to see if a specific
wallet address is a minter of the collection.

## New Mapping and Array
The investors mapping stores a true/false value for a specific address.

The investorList is an array which stores the addresses of minters.

## mintImage() Changes

Minting will check the msg.sender address against the investors
mapping. If the mapping returns false msg.sender will be added to the
mapping with a value of true and the address will be saved in the
investorList. This will be skipped in the address is already an
investor.

## Investor Functions Added
Two new functions added in this section.

The isInvestor() takes an address and returns true if it is an
investor, otherwise returns negative.

The printInvestorList() function returns all addresses stored in the
investorList.

## Documentation Changes
The Statistics, Modification History and Task List were updated for
the current version.