Update to 0.4.0 alpha
# Version 0.4.0 alpha Changes
Very happy to release RenneLaneCollection.sol version 0.4.0 alpha.

This version changes how funds are withdrawn from the contact. Due
to security isses funds are no longer dispersed automatically when
an image is minted, instead the owner must call a new function which
will payout any Ether owed to the artists and send the remaining
balance to the project wallet. Other changes documented below.

## Pragma
Changed to be exactly 0.8.14 for increased security.

## Structs, Variables and Mappings

### Image Struct
The payoutAddress and royalAddress properties were removed and
replaced with an artistId (1-5).

### Artist Struct
This new struct contains the directAddress, royaltyAddress and a
int256 balance to track the funds owed to each artist.

### projectWalletAddress Variable
Added to store wallet for fund dispersal.

### artist Mapping
New mapping to track each artist's addresses and balance using the
above struct. This mapping is initiatialzed in the construtor.

### imageGallery Mapping
Initialization in constructor adjusted for new variables.

## Functions

### mintImage() Changes
This function was adjusted to assign balances to artist mapping.
Automatic payment splitting was removed.

### Assistant Functions
Moved to Admninistrative Functions at the end of the contract. Minor
changes to reflect new data discussed above.

### Administrative Functions

### checkArtistBalances() Added
Takes a uint256 _artistID (1-5) and returns the balance owed to the
artist.

### payArtist() Added
This internal function can only be called by the contact itself, and
is called within the payoutFunds() function. It checks to see if
that artist is owed money and then transfers that amount to that
artists direct wallet. The artists balance is then set to 0.

### payoutFunds() Added
This function will call the payArtist() function for each artist and
then send the remaining funds to the Project Wallet. This function
will revert if there is no money in the contract balance. It will
return True if the total of all balances (artists and contract
balance) is equal to 0.

## Documentation Changes
The Statistics, Modification History and Task List were updated for
the current version.

## Documentation
Documentation overall was converted to better meet NatSpec
standards, specifically for each function. This is something that
will be improved on in future iterations.

### README.md
The readme file was updated to reflect this release.