Add conmpleted ReneeCoins Contract
# ReneeCoins version 0.1.0 alpha

This is a bespoke ERC-20 Contract that provides the social currency,
Renee Coins, which grants benefits to investors and artists of The
Renee Lane Collection.

This contract acts as the ledger for the Renee Coins and allows for
Renee Coins to be created, tracked, and transferred or traded
between wallets. It uses the OpenZeppelin contracts as a back bone
with extension allowing a maximum cap to be specified, and for the
owner to destroy or burn their tokens. The OpenZepplin Ownable
extension is used to set an Owner for the contract and ensure
certain functions are ONLY callable by the Owner.

## Token Cap
The total amount of Renee Coins is capped a 2,000,000 in the
constructor using the ERC20Capped Extension from OpenZeppelin.
There can never be more than 2 million coins existence. Coins that
are burnt can be re-minted up to the cap.

## createCoins() Function

The createCoins() function serves as the minting function for this
contract. It can be called only by the owner and allows them to mint
as many Renee Coins as they like, up to the cap. These function will
ONLY create coins in the Owner's wallet. The airdropCoins can be
used to send coins to another wallet address.

## airdropCoins() Function
The airdropCoins() function allows the Owner to create new coins and
immediately send them to a specified wallet address. This is
significantly cheaper than minting and transferring the coins
separately. This function can ONLY be called by the Owner.