Update to 0.0.0 beta
# Version 0.0.0 beta Changes
Very happy to release the 0.0.0 beta version of The Renee Lane
Collection. The initial feature development of these contracts has completed
and it is being moved to a testing phase on the Rinkeby Testnet.

Updates are listed below for
both contracts. The next release will be the initial beta release where both
contracts will be up for user and integration testing prior to the full
launch.

# ReneeCoins.sol
This is the contract for ERC-20 Renee Coins social currency.

### General Code Changes
Code changes performed for this release were only for updating hardcoded
addresses built into the code. No feature changes, or other refactoring was
done between version 0.5.0 alpha and version 0.0.0 beta.

### License

The MIT License (MIT) is unchanged.

## General Documentation Changes

No major changes to documentation for this release. Comments comply is NatSpec
standard and are written to explain the code in the simplest terms so curious
users can understand what is going on.

### Documentation

Version updated

### Modification History

Modification history from alpha version was removed.

### Statistics

Statistics updated.

### Tasks

The task list was updated for the beta testing phase.

## Resources

### Pragma

Still using solidity version 0.8.15.

### Libraries

Imports for this contract are unchanged.

Contract Imports:

-   OpenZeppelin's ERC20.sol
-   OpenZeppelin's ERC20Burnable.sol
-   OpenZeppelin's ERC20BCapped.sol
-   OpenZeppelin's Ownable.sol

## Contract

The contract section is unchanged.

## Data Structures
There are no data structures for this contract.

## Constructor
No changes to the constructor in this release.

## Functions
Functions have been reorganized to follow the solidity Style Guide and
are now ordered by visibility.

This contract contains the following:

- External Functions
    - None

- OnlyOwner Functions
    - createCoins()
    - airdropCoins()

- Public Functions
    - decimals()

- Internal Functions
    - _mint()

## External Functions

## OnlyOwner Functions

### createCoins() Function
When this contract is initially deployed, there are 0 Renee Coins in
existence. New Renee Coins must be 'minted' in order to create them.
createCoins() serves as the mintingfunction for this contract. It can be
called only by the owner and allows them to mint as many Renee Coins as they
like, up to the cap. This function will ONLY create coins in the Owner's
wallet. The airdropCoins() function can be used to mint coins directly to
someone else's wallet.

### airdropCoins() Function

The airdropCoins() function allows the Owner to mint new Renee Coins directly
into a specified wallet address. This is significantly cheaper than minting
and transferring the coins separately.

This function can ONLY be called by the Owner.

## Public Functions

### decimals() Function
The decimals() function replaces the inherited functions from the ERC20 and
ERC20Capped OpenZeppelin Contracts. For Renee Coins there are 0 decimal
places, these tokens cannot be broken down in smaller increments. This
function is required by the ERC-20 standard.

## Internal Functions

### _mint() Function
The _mint() function replaces the inherited functions from the ERC20 and
ERC20Capped Openzepplin Contracts. This is an internal function which can only
be called by the contract itself and is used inside other functions such as
createCoins() and airdropCoins().

# Items Removed
No items were moved for this release

# README.md
The readme file was updated to reflect this release. 