Update to 0.0.0 beta
# Version 0.0.0 beta Changes
Very happy to release the 0.0.0 beta version of The Renee Lane
Collection. The initial feature development of these contracts has completed
and it is being moved to a testing phase on the Rinkeby Testnet.

Updates are listed below for  both contracts. The next release will be the
initial beta release where both contracts will be up for user and integration
testing prior to the full launch.

# Python Scripts

Begun creating Python scripts to help manage the contract.

# README.md
The readme file was updated to reflect this release. 